# Memberlist/Ring: do not normalise tokens on subsequent Merge().