<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

**What this PR does**:

Improves performance of `ring.Desc.Merge()`.

Memberlist uses the Merge() function on ring.Desc such that it is only ever
called subsequently on the result of itself. Therefore, because the Merge()
function always returns Desc in the normalised form, it does not have to
normalise again before doing another Merge().

Unit tests are fixed up and extended to reflect the precondition that the LHS
must be normalised, and the RHS does not have to be normalised.

I suspect it is also possible to remove normalisation on the incoming change
(the Desc being merged into the current Desc), but it would be safest to test
these changes incrementally. (Open to opinions as to omitting both now)

Performance difference:
```
name                             old time/op    new time/op    delta
MemberlistReceiveWithRingDesc-6     414µs ± 5%      31µs ± 7%  -92.49%  (p=0.000 n=10+10)

name                             old alloc/op   new alloc/op   delta
MemberlistReceiveWithRingDesc-6    22.7kB ± 0%     8.3kB ± 0%  -63.54%  (p=0.000 n=9+10)

name                             old allocs/op  new allocs/op  delta
MemberlistReceiveWithRingDesc-6       655 ± 0%        55 ± 0%  -91.60%  (p=0.000 n=10+10)
```


**Which issue(s) this PR fixes**:

Fixes n/a

**Checklist**
- [x] Tests updated
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`