@stevesg 2021-12-09
> I think we should be careful about removing normalization of incoming message (RHS). While our code produces messages that are normalized already (eg. sorted tokens), there may be clients using dskit where this doesn't apply. If we want to guarantee that value stored in KV store is normalized, I think we cannot drop normalization of incoming messages.

Good point. I suspect that it will be safest, and necessary, to always normalise whatever state we receive from CAS() operations, though we could possibly avoid it for broadcasted changes which were generated by Merge()? Could explore it separately if it comes up in profiling.