@pstibrany 2021-12-09
In a private discussion @pracucci suggested having an option to enable/disable normalization of incoming value. We know we can disable it in our code, but it can default to enabled just in case.